move `_maybe_print_use_warning` into delta_generator.py
