+++ lib/tests/streamlit/streamlit_test.py
@@ -82,28 +82,6 @@ class StreamlitTest(unittest.TestCase):
         with self.assertRaises(StreamlitAPIException):
             st.set_option("server.enableCORS", False)
 
-    @patch("streamlit.runtime.Runtime.exists", MagicMock(return_value=False))
-    def test_run_warning_presence(self):
-        """Using Streamlit without `streamlit run` produces a warning."""
-        with self.assertLogs(level=logging.WARNING) as logs:
-            st._use_warning_has_been_displayed = False
-            st.write("Using delta generator")
-            output = "".join(logs.output)
-            # Warning produced exactly once
-            self.assertEqual(len(re.findall(r"streamlit run", output)), 1)
-
-    @patch("streamlit.runtime.Runtime.exists", MagicMock(return_value=True))
-    def test_run_warning_absence(self):
-        """Using Streamlit through the CLI results in a Runtime being instantiated,
-        so it produces no usage warning."""
-        with self.assertLogs(level=logging.WARNING) as logs:
-            st._use_warning_has_been_displayed = False
-            st.write("Using delta generator")
-            # assertLogs is being used as a context manager, but it also checks
-            # that some log output was captured, so we have to let it capture something
-            get_logger("root").warning("irrelevant warning so assertLogs passes")
-            self.assertNotRegex("".join(logs.output), r"streamlit run")
-
 
 class StreamlitAPITest(testutil.DeltaGeneratorTestCase):
     """Test Public Streamlit Public APIs."""
